refactor(widget): separate OfferCard from Home screen for modularity
