+++ lib/features/home/presentation/screens/home_screen.dart
@@ -1,11 +1,10 @@
 import 'package:auto_route/annotations.dart';
-import 'package:eatsy/common/constants/app_colors.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
 import '../../../../common/widgets/background_pattern_base_screen.dart';
 import '../widgets/header_bar.dart';
+import '../widgets/offer_card.dart';
 
 @RoutePage()
 class HomeScreen extends StatelessWidget {
@@ -26,65 +25,5 @@ class HomeScreen extends StatelessWidget {
         ),
       ),
     ));
-  }
-}
-
-class OfferCard extends StatelessWidget {
-  const OfferCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 180,
-      width: double.maxFinite,
-      child: DecoratedBox(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(14),
-          gradient: AppColors.greenGradient,
-        ),
-        child: Stack(
-          fit: StackFit.expand,
-          children: [
-            SvgPicture.asset('assets/icons/card_mask.svg'),
-            Row(
-              children: [
-                SizedBox(
-                  height: 160,
-                  child: Image.asset(
-                    'assets/images/ice_cream_offer.png',
-                    fit: BoxFit.fitHeight,
-                  ),
-                ),
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.end,
-                  children: [
-                    const Text(
-                      "Special Deal For \nOctober",
-                      style: TextStyle(
-                        fontSize: 20,
-                        color: Colors.white,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                    MaterialButton(
-                      color: Colors.white,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      onPressed: () {},
-                      child: const Text(
-                        "Buy Now",
-                        style: TextStyle(color: AppColors.green),
-                      ),
-                    )
-                  ],
-                )
-              ],
-            )
-          ],
-        ),
-      ),
-    );
   }
 }
